% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wip_allmeans.R
\name{allmeans}
\alias{allmeans}
\title{Title}
\usage{
allmeans(dataset, variables)
}
\arguments{
\item{dataset}{Dataset}

\item{variables}{Vector/list of variables of interest, same name as columns in dataset}
}
\description{
Title
}
