% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemeanfee.R
\name{comparemeanfee}
\alias{comparemeanfee}
\title{Compare mean fees of two datasets}
\usage{
comparemeanfee(dataone, datatwo)
}
\arguments{
\item{dataone}{First dataset, must contain column "Fee"}

\item{datatwo}{Second dataset, must contain column "Fee"}
}
\value{
Returns a paired t-test for the Fee numeric vectors in each dataset
}
\description{
Compare mean fees of two datasets
}
\examples{
comparemeanfee(engdata,gerdata)

}
