% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasubsets.R
\name{datasubsets}
\alias{datasubsets}
\title{Create data subsets for levels of chosen variable}
\usage{
datasubsets(dataset, variable = "Year")
}
\arguments{
\item{dataset}{Dataset, acquired from storedata}

\item{variable}{Categorical variable whose levels will define the subsets}
}
\value{
Returns the subsets of the dataset as a vector
}
\description{
Create data subsets for levels of chosen variable
}
\examples{
datasubsets(engdata)
datasubsets(gerdata,"Team")

}
