% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultsettings.R
\name{defaultsettings}
\alias{defaultsettings}
\title{Set default settings that author defined}
\usage{
defaultsettings()
}
\value{
Establishes settings that author used during creation of package
}
\description{
Set default settings that author defined
}
\examples{
defaultsettings()

}
