% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{engdata}
\alias{engdata}
\title{Formatted excel data using storedata() used for functions}
\format{
A data frame with rows corresponding to players and 10 variables
\describe{
\item{Team}{Current Team}
\item{Name}{Player Name}
\item{Age}{Age At Time of Transfer}
\item{Position}{Player Position}
\item{Fee}{Fee in Euros}
\item{From}{Previous Team}
\item{Year}{Transfer Window of Acquisition}
\item{National}{National Team}
\item{Capped}{Senior National Appearance Prior to Transfer Y/N}
\item{Remarks}{Empty Unless the Player Is On Loan}
}
}
\source{
{Created in-house using information from transfermarkt.us}
}
\usage{
engdata
}
\description{
Contains information for various league teams as of the Summer '23 window
}
\examples{
data(engdata)
}
\keyword{datasets}
