% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessgreaterthanfee.R
\name{lessgreaterthanfee}
\alias{lessgreaterthanfee}
\title{Create subset of data with fees less/greater than inputted fee value}
\usage{
lessgreaterthanfee(dataset, fee = 5010000, less = FALSE)
}
\arguments{
\item{dataset}{Dataset, acquired from storedata}

\item{fee}{Fee value that defines cutoff for data, defaults to peak average transfer (2019/20)}

\item{less}{Looking for observations less than inputted fee (T/F)}
}
\value{
Return dataset based on inputted parameters
}
\description{
Create subset of data with fees less/greater than inputted fee value
}
\examples{
lessgreaterthanfee(engdata)
lessgreaterthanfee(engdata,fee=10000000,less=TRUE)

}
