% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfeebyage.R
\name{meanfeebyage}
\alias{meanfeebyage}
\title{Plot graph of mean fee by Age}
\usage{
meanfeebyage(dataset)
}
\arguments{
\item{dataset}{Dataset, acquired from storedata}
}
\value{
Creates and displays ggplot2 graph of mean fee by age
}
\description{
Plot graph of mean fee by Age
}
\examples{
meanfeebyage(engdata)

}
