% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfeebypos.R
\name{meanfeebypos}
\alias{meanfeebypos}
\title{Plot graph of mean fee by Position}
\usage{
meanfeebypos(dataset)
}
\arguments{
\item{dataset}{Dataset, acquired from storedata}
}
\value{
Creates and displays ggplot2 graph of mean fee by position
}
\description{
Plot graph of mean fee by Position
}
\examples{
meanfeebypos(engdata)

}
