% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfeebyyear.R
\name{meanfeebyyear}
\alias{meanfeebyyear}
\title{Plot graph of mean fee by Year}
\usage{
meanfeebyyear(dataset)
}
\arguments{
\item{dataset}{Dataset, acquired from storedata}
}
\value{
Creates and displays ggplot2 graph of mean fee by year
}
\description{
Plot graph of mean fee by Year
}
\examples{
meanfeebyyear(engdata)

}
