% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfeegraph.R
\name{meanfeegraph}
\alias{meanfeegraph}
\title{Create ggplot2 graph of mean fee by level of chosen categorical variable}
\usage{
meanfeegraph(dataset, variable = "Year")
}
\arguments{
\item{dataset}{Dataset, acquired from storedata}

\item{variable}{Categorical explanatory variable}
}
\value{
Returns ggplot2 graphic of mean fee by explanatory levels
}
\description{
Create ggplot2 graph of mean fee by level of chosen categorical variable
}
\examples{
meanfeegraph(engdata)
meanfeegraph(engdata,"Age")
}
