% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentlocal.R
\name{percentlocal}
\alias{percentlocal}
\title{Find proportion of local players (player nationality same as league)}
\usage{
percentlocal(dataset, country)
}
\arguments{
\item{dataset}{Dataset, must contain column "National"}

\item{country}{Country defined as local}
}
\value{
Returns the proportion of players whose nationality matches the locality
}
\description{
Find proportion of local players (player nationality same as league)
}
\examples{
percentlocal(engdata,"England")
}
