% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedata.R
\name{storedata}
\alias{storedata}
\title{Store excel data in an R DataFrame}
\usage{
storedata(filepath, cellrange)
}
\arguments{
\item{filepath}{File path that points to the excel spreadsheet}

\item{cellrange}{Range of cells in format 'SheetName!TopLeft:BottomRight'}
}
\value{
Returns the excel sheet as an R data frame
}
\description{
Store excel data in an R DataFrame
}
\examples{
storedata(filepath="the/file/path",cellrange="England!A1:J500")

}
